RFC 9022: location Spalte in Suchtabellen, bugfixes Ulf, Rolle AIR Location Data Maintenance
git-svn-id: http://10.205.116.236/svn/AIR@350 d2170373-ae6f-44b7-9150-2eec3990abfc

+++ branches/2.00/src/com/bayerbbs/applrepos/domain/ServiceContract.java
@@ -11,77 +11,33 @@ import javax.persistence.Transient;
 @Entity
 @Table(name = "SERVICE_CONTRACT")
 public class ServiceContract extends DeletableRevisionInfo implements Serializable {
-
 	private static final long serialVersionUID = -924659591525578344L;
 
-	private Long serviceContractId;
+	private Integer serviceContractId;
 	private String serviceContract;
 
-	
-	// ------------------------------------------------------
-	// -
-	// ------------------------------------------------------
-
-	/**
-	 * Creates a new instance.
-	 */
 	public ServiceContract() {
 	}
 
-	// ------------------------------------------------------
-	// hibernate get / set
-	// ------------------------------------------------------
-
-	/**
-	 * Returns the value of the field {@link #id}.
-	 * 
-	 * @return Value of the {@link #id} field.
-	 */
 	@Transient
-	public Long getId() {
+	public Integer getId() {
 		return getServiceContractId();
 	}
 
-	/**
-	 * Returns the value of the field {@link #serviceContractId}.
-	 * 
-	 * @return Value of the {@link #serviceContractId} field.
-	 */
 	@Id
 	@Column(name = "SERVICE_CONTRACT_ID")
-	public Long getServiceContractId() {
+	public Integer getServiceContractId() {
 		return serviceContractId;
 	}
-
-	/**
-	 * Sets the value of the {@link #serviceContractId} field.
-	 * 
-	 * @param serviceContractId
-	 *            The value to set.
-	 */
-	public void setServiceContractId(Long serviceContractId) {
+	public void setServiceContractId(Integer serviceContractId) {
 		this.serviceContractId = serviceContractId;
 	}
 	
-	/**
-	 * Returns the value of the field {@link #serviceContract}.
-	 * 
-	 * @return Value of the {@link #serviceContract} field.
-	 */
 	@Column(name = "SERVICE_CONTRACT")
 	public String getServiceContract() {
 		return serviceContract;
 	}
-
-	/**
-	 * Sets the value of the {@link #serviceContract} field.
-	 * 
-	 * @param serviceContract
-	 *            The value to set.
-	 */
 	public void setServiceContract(String serviceContract) {
 		this.serviceContract = serviceContract;
 	}
-	
-
 }
